Small KAMT improvements (#1357)
* Improve wording and logic around extension lengths

Specifically:

- Rename "consumed" to "length", because this is actually the length.
- Improve `HashBits` to store a length internally. Otherwise, we could
  end up reading past the end when constructing a `HashBits` from an
  `Extension`. This isn't actually possible due to how we use it, but I
  don't want to rely on it.

* KAMT: use u32s for lengths

There's really no benefit to using a u8 here due to alignment. It'll
probably just be more expensive as we cast back and for math
operations (e.g., wasm doesn't have u8 math operations).

This main benefit is that there's no chance of randomly overflowing.

* fix: remove docs talking about why we're using a u8 for a length

* fix: avoid allocating a temporary vector

We can just pass an array here, and we don't need to filter either. This
probably won't make too much of a performance difference, but it makes
it easier to understand (for me, at least).